+++ scm-core/src/main/java/sonia/scm/security/SessionId.java
@@ -24,15 +24,24 @@
     
 package sonia.scm.security;
 
+import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Preconditions;
 import com.google.common.base.Strings;
+import lombok.EqualsAndHashCode;
+import sonia.scm.util.HttpUtil;
 
-import java.util.Objects;
+import javax.servlet.http.HttpServletRequest;
+import java.io.Serializable;
+import java.util.Optional;
 
 /**
  * Client side session id.
  */
-public final class SessionId {
+@EqualsAndHashCode
+public final class SessionId implements Serializable {
+
+  @VisibleForTesting
+  public static final String PARAMETER = "X-SCM-Session-ID";
 
   private final String value;
 
@@ -41,21 +50,12 @@ public final class SessionId {
   }
 
   @Override
-  public boolean equals(Object o) {
-    if (this == o) return true;
-    if (o == null || getClass() != o.getClass()) return false;
-    SessionId sessionID = (SessionId) o;
-    return Objects.equals(value, sessionID.value);
-  }
-
-  @Override
-  public int hashCode() {
-    return Objects.hash(value);
-  }
-
-  @Override
   public String toString() {
     return value;
+  }
+
+  public static Optional<SessionId> from(HttpServletRequest request) {
+    return HttpUtil.getHeaderOrGetParameter(request, PARAMETER).map(SessionId::valueOf);
   }
 
   public static SessionId valueOf(String value) {