+++ scm-plugins/scm-git-plugin/src/test/java/sonia/scm/api/v2/resources/GitConfigResourceTest.java
@@ -248,7 +248,7 @@ public class GitConfigResourceTest {
     assertThat(configurationStoreCaptor.getValue())
       .isInstanceOfSatisfying(GitRepositoryConfig.class, x -> { })
       .extracting("defaultBranch")
-      .containsExactly("new");
+      .isEqualTo("new");
   }
 
   private MockHttpResponse get() throws URISyntaxException {