+++ scm-webapp/src/test/java/sonia/scm/api/v2/resources/DiffResultToDiffResultDtoMapperTest.java
@@ -79,7 +79,7 @@ class DiffResultToDiffResultDtoMapperTest {
       .isPresent()
       .get()
       .extracting("href")
-      .contains("/scm/api/v2/repositories/space/X/diff/123/parsed");
+      .isEqualTo("/scm/api/v2/repositories/space/X/diff/123/parsed");
   }
 
   @Test
@@ -91,7 +91,7 @@ class DiffResultToDiffResultDtoMapperTest {
       .isPresent()
       .get()
       .extracting("href")
-      .contains("/scm/api/v2/repositories/space/X/incoming/feature%2Fsome/master/diff/parsed");
+      .isEqualTo("/scm/api/v2/repositories/space/X/incoming/feature%2Fsome/master/diff/parsed");
   }
 
   private DiffResult createResult() {